requirements.txt
fastapi
pydantic
uvicorn
torch
transformers
pillow
requests
numpy